requirements.txt
flask
flask_cors
Flask-SQLAlchemy
psycopg2-binary